Fix docker container tests
Previously, the docker container always had a postgres user, which was
then used to create the database and superuser that was specified.
The tests that don't use the containers also emulated this, for the same
permission behaviour. Now, we make it work like that again. We could
remove it, but it gives us an additional user to grant permissions to
for testing.